Handle cns files with missing probe counts
We ended up with a .cns file having a '_' for `probes` which
caused issues on export to VCF since pandas converted the `probes`
column to a string instead of the expected integer. This is a band-aid
that skips rows without probes and assures other rows are integers
before feeding into string formatting.

This also bumps the version to 0.7.2.dev0 since .dev0 versions sort
before full versions (0.7.1.dev0 < 0.7.1)