export nexus-ogt: mirror median BAF the same direction as most input BAFs
If multiple SNVs fall in a bin, we summarize their BAFs by taking the median of
the mirrored BAFs. But mirroring can go above or below .5. To reduce asymmetry
in the output visualizations, mirror the BAFs in the same direction as the
majority of input SNV BAFs; do this by testing the median vs. .5 to break ties
for an even number of input SNVs.