test: add a VCF of SNVs from mixed 1000 Genomes samples NA12878, NA12882
Simulates a tumor with 75% purity where son-specific calls are somatic SNVs.
In the "tumor-normal" pair used for calling, the "normal" sample is 1000 Genomes
mother NA12878, downsampled to 200x, and the "tumor" is a mix of of 25% mother
and 75% son NA12882, downsampled to 300x. Physical reference samples
(probably from Horizon?) were sequenced with a hybrid capture protocol.