segment: Handle numeric sample IDs in cbs, flasso methods' output
When CNVkit read the SEG output from R, integer sample names were parsed as
integers, so comparing the string sample ID from the input filename to the
integer sample ID parsed from SEG would not find a match. A blank CNA with
minimal columns would be created as a fallback, and an error would finally occur
when combining dataframes with unequal numbers of columns. Spooky.

Now, tabio.seg.parse_seg forces the parsed sample_id to be string, and
tabio.seg.read_seg raises an error if a sample ID or index was specified but is
not found.