tabio.vcfio: Extract locus depths from bcftools VCF (closes #146)
In this case, the DP field is in the INFO column instead of the sample column.
The allele depths are not listed anywhere (AFAICT), so the result will have
valid 'depth' values, but 'allele_count' and 'allele_freq' will be all 0.0. Most
CNVkit functions will "work" without crashing, but BAF calculations and plots
will be fairly pointless.