autobin: handle an edge case in 'hybrid' mode (chapmanb/bcbio-nextgen#1799)
The reported bug is handled by the previous commit. This one handles the
possibility that all of the sequencing-accessible regions on a chromosome are
covered by targets, i.e. access - targets = {}. This would be presumably WGS or
user error in supplying the same BED file for access and targets, or switching
the two. But this at least ensures the command doesn't crash.