import-picard: Don't accept directory as an argument (was deprecated)
Be a little more flexible in filenames accepted: instead of requiring
input files to be named '*.targetcoverage.???' or
'*.antitargetcoverage.???', strip the full suffix and default to
'targetcoverage.cnn' output suffix, or 'antitargetcoverage.cnn' if input
filename contains 'antitarget'. Works the same for filenames following
the earlier convention, but now pretty safe for amplicon targets with
arbitrary filenames, and slightly less spooky.