export cdt: Fix swapped columns in output; add "EWIEGHT" row for compat
Build dataframes from a list of columns & value arrays instead of a
dict, which was ignoring order and causing column values to disagree
with the header. (Reported by HST)

Also add a dummy EWEIGHT row, per online specs, to ensure Java TreeView
starts reading data rows from the correct row. Not sure if this was ever
a problem in any JTV versions.