call,segment: Always add a "baf" column if a VCF is given (closes #224)
These code blocks would be skipped if the VCF did not contain a
recognizable paired normal sample.
(1) Require only the tumor sample's allele frequencies be present for
the calculation (but still also use the normal's if found);
(2) if even the tumor AF is missing, warn & return a column of NaN

This should still avoid the crash seen in #218.