scatter: fix a regression in "Selection" behavior with --by-bins
The "Selection" box (highlighted non-genic region) should only be shown if there
are no genes in the specified genomic region. With --by-bins, this box was being
shown even if there were genes in the region, due to inconsistency between the
outputs of `skgenome.rangelabel.unpack_range()` and
`cnvlib.plots.translate_region_to_bins()`.