segmetrics, genemetrics: Make list-type args default to [] (#325)
The stats options in segmetrics rely on argparse to build a list of the
names the statistics to be calculated. This should default to an empty
list if no stats (in a category) are specified, but it was defaulting to
None.

This commit fixes the issue in segmetrics, where the crash was
observed. While the genemetrics command has the same options available,
they're not implemented internally yet so the command didn't crash --
but the same solution is needed there, too.