iter_slices: Respect keep_empty the same way as by_ranges
Where the 'destination' table has chromosomes that the 'source' table
does not, emit a number of empty slices/indices matching the number of
rows on that chromosome in the 'destination' table. This ensures that,
with keep_empty=True, the number of yielded elements matches the number
of rows in 'destination'.

(Without this, into_ranges could create an output array that was too
short, crashing the 'target' command's unit test.)