segmetrics: Implement & expose --smooth-bootstrap option
Settled on an approach where KDE bandwidth is based on each bin's
weight, which indicates a decent estimate of that bin's log2 SD.
This is consistent with the 'bintest' approach, so 'p_bintest' and 'ci'
should reflect the same inferred distribution for a single-bin segment.

Literature says the smoothing distribution is not hugely important, but
there are some risks to over-smoothing for larger sample sizes, so we
use a loose interpretation of Silverman's Rule to reduce the bandwidth
as the number of bins in a segment increases (k^-1/4).

Update unit tests and test Makefile to include the --t-test and
--smooth-bootstrap options, and also 'bintest' which I skipped earlier.