Fix smoothing bug when signal is shorter than a window
In Savitzky-Golay smoothing, for certain combinations of parameters, a
(wing-padded) signal length can turn out to be shorter than the
smoothing window requested, which is mathematically impossible to
compute and causes an index mismatch exception downstream.

This change avoids this by checking for this condition and reducing the
smoothing window length and, if necessary, also the polynomial order.
This ensures (compared to e.g. returning the original signal) that at
least some sensible smoothing is still being performed.