Improve error message when coordinate set contains duplicate values
Closes #637.

Previously, when the coordinate set contains lots of duplicate values
(on the order of thousands and more), the exception would format and
dump all of them to stderr, obscuring the initial error message.

Now, only the number of duplicated regions is output, with up to ten
first regions, making the error more readable in all cases.