fix: Choose a smoothing window fraction based on the data size (#379)
Previously it was fixed at 0.1; now it's 1/sqrt(length), down to a limit of 0.01.

Behavior is unchanged for a vector length of 100. For smaller vectors, e.g.
down to ~20 as could be seen in the test file amplicon.cnr, the smoothing window
is wider (~.22, or 4 data points), and for larger vectors shrinks down to a
limit of 0.01 for vector length >= 10000 (for >=100 data points in the window).

Compare to GC-binning functionality in Picard tools and GATK, where a similar
adjustment is made at each percentile rank, equivalent to a fixed window
fraction of .01 but without rolling/smoothing.

This should correct biases better at the extremes of the GC range in particular
(when there are enough data points), where previously the fitted line would
flatten out at the edges due to the wide smoothing window.