Migrate to pyproject.py backed by setuptools
The standalone executables in `scripts/*` are no longer installed
automatically. PyPA does not seem to encourage that anymore and
setuptools doesn't have easy support for it -- it needs a shim function
for each script to install.

Added a shim to install cnvkit.py but not the other scripts.

Also, display help text automatically when cnvkit.py is run without
arguments.