Consider PAR1/2 on Chr X as autosomal (#789)
Add the optional feature to treat the PAR1/2 on chromosome X as autosomal, i.e. when requesting autosomal regions these regions are included and on the other hand, when requesting "chromosome X", they are excluded. For chr Y, the PAR1/2 regions are excluded when requesting "chromosome Y" since they cannot be covered if all reads map to PAR1/2 on X. Whenever the code treats chromosome X, one doesn't have to reason about whether PAR1/2 is included or not. No differentiation is done based on the sex of the sample.

Specifically:
* Add new command line argument --diploid-parx-genome which can be set to grch37 or grch38 in order to treat PAR1/2 GRCh37/38 regions on chromosome X as autosomal.
* Reimplement the method GenomicArray.autosomes() in CopyNumArray.autosomes() to be able to transparently include/exclude PAR1/2 on demand. Conversely, wherever chromosome X regions have been requested (ex: (cnarr.chromosome == cnarr._chr_x_label)), this has been replaced by a call to CopyNumArray.chr_x_filter() which includes/excludes PAR1/2 as needed. Same for chromosome Y.
* Add read_ga similar to read_cna() and use where appropriate (=GenomicArray tests)
* Add CopyNumArray methods `chr_x_filter`, `chr_y_filter`, and `pary_filter`. When using cnvlib, you should now use `chrx = self[self.chrx_filter()]` instead of `self[self.chromosome == self.chr_x_label]`.


See also: #784 "Optionally consider PAR1/2 as diploid in male samples due to reference genome masking" -- with this changeset, if enabled, the regions for chrX will be split even for female samples (where it doesn't make sense, as opposed to for male samples).


cnvlib.cnary: 
* Replace `_chr_x_label` with `chr_x_label + chr_x_filter()`
* Refactor abs/ref/exp copy calculations, absolute_expect() & absolute_reference().
* Fix bug: access the same array by 2 variables
* Allow autosomes()'s also parameter to be a pd Series.
* s/get_raw_dataframe/get_copied_raw_dataframe/g
* Use chr x/y label instead of hardcoded names.

cnvlib.commands:
* Forward parameter from cli as needed.
* Add --procs argument to reference.

test:
* Add --diploid-parx-genome test.
* Make gary tests only use GenomicArray.
* Add and refactor tests.
* Add missing test file.
* Test clonal, too.
* s/read/read_ga/g in tests.
* Improve test data.

Other fixes:
* Fix a bug in call.absolute_dataframe().
* Fix deprecation warning.
* Fix export_vcf command/method.
* Fix metrics.
* Allow the .fai file to be older than the .fa file.
* Remove option to rebuild fasta index.
* Pin pomegranate version to exactly 0.14.8